Add definition for CPU::Reset function
The Reset function initializes the CPU's registers and memory addresses to their default states. This includes setting the Accumulator, X, and Y registers to 0, the StackPointer to 0xFD, and the StatusRegister to 0x00 with the unused flag set. It also sets the ProgramCounter to the address stored at 0xFFFC, initializes the RelativeAddress, AbsoluteAddress, and FetchedData to 0, and sets the CyclesLeft to 8.

Signed-off-by: Calisto Abel Mathias <calisto.mathias.25@gmail.com>